part02: Complete exercise 2.5
