+++ part02/courseinfo/src/App.js
@@ -1,55 +1,4 @@
-const Header = ({ course }) => {
-  return <h2>{course}</h2>
-}
-
-const Part = ({ part, exercises }) => {
-  return (
-    <p>
-      {part} {exercises}
-    </p>
-  )
-}
-
-const Content = ({ parts }) => {
-  return (
-    <div>
-      {parts.map((part) => (
-        <Part key={part.id} part={part.name} exercises={part.exercises} />
-      ))}
-    </div>
-  )
-}
-
-const Total = ({ parts }) => {
-  const exercises = parts
-    .map((part) => part.exercises)
-    .reduce((a, b) => {
-      return a + b
-    })
-  return <b>total of {exercises} exercises</b>
-}
-
-const Course = ({ course }) => {
-  const { name, parts } = course
-  return (
-    <div>
-      <Header course={name} />
-      <Content parts={parts} />
-      <Total parts={parts} />
-    </div>
-  )
-}
-
-const Courses = ({ courses }) => {
-  return (
-    <div>
-      <h1>Web development curriculum</h1>
-      {courses.map((course) => (
-        <Course key={course.id} course={course} />
-      ))}
-    </div>
-  )
-}
+import Courses from './Courses'
 
 const App = () => {
   const courses = [
