+++ part02/courseinfo/src/Course.js
@@ -1,0 +1,43 @@
+const Header = ({ course }) => {
+  return <h2>{course}</h2>
+}
+
+const Part = ({ part, exercises }) => {
+  return (
+    <p>
+      {part} {exercises}
+    </p>
+  )
+}
+
+const Content = ({ parts }) => {
+  return (
+    <div>
+      {parts.map((part) => (
+        <Part key={part.id} part={part.name} exercises={part.exercises} />
+      ))}
+    </div>
+  )
+}
+
+const Total = ({ parts }) => {
+  const exercises = parts
+    .map((part) => part.exercises)
+    .reduce((a, b) => {
+      return a + b
+    })
+  return <b>total of {exercises} exercises</b>
+}
+
+const Course = ({ course }) => {
+  const { name, parts } = course
+  return (
+    <div>
+      <Header course={name} />
+      <Content parts={parts} />
+      <Total parts={parts} />
+    </div>
+  )
+}
+
+export default Course
