+++ part02/courseinfo/src/Courses.js
@@ -1,0 +1,14 @@
+import Course from './Course'
+
+const Courses = ({ courses }) => {
+  return (
+    <div>
+      <h1>Web development curriculum</h1>
+      {courses.map((course) => (
+        <Course key={course.id} course={course} />
+      ))}
+    </div>
+  )
+}
+
+export default Courses
